refactor: JPA 엔티티를 infrastructure 레이어로 이동
