--- src/main/java/ojik/ojikback/domain/enums/AttendanceStatus.java
@@ -1,6 +1,0 @@
-package ojik.ojikback.domain.enums;
-
-public enum AttendanceStatus {
-    CERTIFIED,
-    REVOKED
-}
