--- src/main/java/ojik/ojikback/domain/enums/GameStatus.java
@@ -1,6 +1,0 @@
-package ojik.ojikback.domain.enums;
-
-public enum GameStatus {
-    SCHEDULED,
-    FINISHED
-}
