--- src/main/java/ojik/ojikback/domain/enums/SocialProvider.java
@@ -1,6 +1,0 @@
-package ojik.ojikback.domain.enums;
-
-public enum SocialProvider {
-    KAKAO,
-    NAVER
-}
