+++ src/main/java/ojik/ojikback/infrastructure/repository/entity/enums/AttendanceStatus.java
@@ -1,0 +1,6 @@
+package ojik.ojikback.infrastructure.repository.entity.enums;
+
+public enum AttendanceStatus {
+    CERTIFIED,
+    REVOKED
+}
