+++ src/main/java/ojik/ojikback/infrastructure/repository/entity/enums/GameStatus.java
@@ -1,0 +1,6 @@
+package ojik.ojikback.infrastructure.repository.entity.enums;
+
+public enum GameStatus {
+    SCHEDULED,
+    FINISHED
+}
