+++ src/main/java/ojik/ojikback/infrastructure/repository/entity/enums/SocialProvider.java
@@ -1,0 +1,6 @@
+package ojik.ojikback.infrastructure.repository.entity.enums;
+
+public enum SocialProvider {
+    KAKAO,
+    NAVER
+}
